backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2

pdfplumber==0.11.4
python-docx==1.1.2

spacy==3.7.5
scikit-learn==1.5.1
numpy==1.26.4


